+++ packages/ai-constructs/src/conversation/runtime/bedrock_converse_adapter.test.ts
@@ -14,6 +14,7 @@ import {
   ConverseCommandOutput,
   Message,
   ToolConfiguration,
+  ToolInputSchema,
   ToolResultContentBlock,
 } from '@aws-sdk/client-bedrock-runtime';
 import { ConversationTurnEventToolsProvider } from './event-tools-provider';
@@ -173,7 +174,14 @@ void describe('Bedrock converse adapter', () => {
               toolUse: {
                 toolUseId: randomUUID().toString(),
                 name: additionalTool.name,
-                input: 'additionalToolInput',
+                input: 'additionalToolInput1',
+              },
+            },
+            {
+              toolUse: {
+                toolUseId: randomUUID().toString(),
+                name: additionalTool.name,
+                input: 'additionalToolInput2',
               },
             },
           ],
@@ -194,7 +202,14 @@ void describe('Bedrock converse adapter', () => {
               toolUse: {
                 toolUseId: randomUUID().toString(),
                 name: eventTool.name,
-                input: 'eventToolToolInput',
+                input: 'eventToolToolInput1',
+              },
+            },
+            {
+              toolUse: {
+                toolUseId: randomUUID().toString(),
+                name: eventTool.name,
+                input: 'eventToolToolInput2',
               },
             },
           ],
@@ -252,14 +267,14 @@ void describe('Bedrock converse adapter', () => {
           toolSpec: {
             name: eventTool.name,
             description: eventTool.description,
-            inputSchema: eventTool.inputSchema,
+            inputSchema: eventTool.inputSchema as ToolInputSchema,
           },
         },
         {
           toolSpec: {
             name: additionalTool.name,
             description: additionalTool.description,
-            inputSchema: additionalTool.inputSchema,
+            inputSchema: additionalTool.inputSchema as ToolInputSchema,
           },
         },
       ],
@@ -288,6 +303,10 @@ void describe('Bedrock converse adapter', () => {
       additionalToolUseBedrockResponse.output?.message?.content[0].toolUse
         ?.toolUseId
     );
+    assert.ok(
+      additionalToolUseBedrockResponse.output?.message?.content[1].toolUse
+        ?.toolUseId
+    );
     const expectedBedrockInput2: ConverseCommandInput = {
       messages: [
         ...(messages as Array<Message>),
@@ -304,6 +323,15 @@ void describe('Bedrock converse adapter', () => {
                     .toolUse.toolUseId,
               },
             },
+            {
+              toolResult: {
+                content: [additionalToolOutput],
+                status: 'success',
+                toolUseId:
+                  additionalToolUseBedrockResponse.output?.message.content[1]
+                    .toolUse.toolUseId,
+              },
+            },
           ],
         },
       ],
@@ -315,6 +343,9 @@ void describe('Bedrock converse adapter', () => {
     assert.ok(eventToolUseBedrockResponse.output?.message?.content);
     assert.ok(
       eventToolUseBedrockResponse.output?.message?.content[0].toolUse?.toolUseId
+    );
+    assert.ok(
+      eventToolUseBedrockResponse.output?.message?.content[1].toolUse?.toolUseId
     );
     assert.ok(expectedBedrockInput2.messages);
     const expectedBedrockInput3: ConverseCommandInput = {
@@ -330,6 +361,15 @@ void describe('Bedrock converse adapter', () => {
                 status: 'success',
                 toolUseId:
                   eventToolUseBedrockResponse.output?.message.content[0].toolUse
+                    .toolUseId,
+              },
+            },
+            {
+              toolResult: {
+                content: [eventToolOutput],
+                status: 'success',
+                toolUseId:
+                  eventToolUseBedrockResponse.output?.message.content[1].toolUse
                     .toolUseId,
               },
             },
@@ -689,14 +729,14 @@ void describe('Bedrock converse adapter', () => {
           toolSpec: {
             name: additionalTool.name,
             description: additionalTool.description,
-            inputSchema: additionalTool.inputSchema,
+            inputSchema: additionalTool.inputSchema as ToolInputSchema,
           },
         },
         {
           toolSpec: {
             name: clientTool.name,
             description: clientTool.description,
-            inputSchema: clientTool.inputSchema,
+            inputSchema: clientTool.inputSchema as ToolInputSchema,
           },
         },
       ],
@@ -799,5 +839,44 @@ void describe('Bedrock converse adapter', () => {
         ],
       },
     ]);
+  });
+
+  void it('adds user agent middleware', async () => {
+    const event: ConversationTurnEvent = {
+      ...commonEvent,
+    };
+
+    event.request.headers['x-amz-user-agent'] = 'testUserAgent';
+
+    const bedrockClient = new BedrockRuntimeClient();
+    const addMiddlewareMock = mock.method(bedrockClient.middlewareStack, 'add');
+
+    new BedrockConverseAdapter(
+      event,
+      [],
+      bedrockClient,
+      undefined,
+      messageHistoryRetriever
+    );
+
+    assert.strictEqual(addMiddlewareMock.mock.calls.length, 1);
+    const middlewareHandler = addMiddlewareMock.mock.calls[0].arguments[0];
+    const options = addMiddlewareMock.mock.calls[0].arguments[1];
+    assert.strictEqual(options.name, 'amplify-user-agent-injector');
+    const args: {
+      request: {
+        headers: Record<string, string>;
+      };
+    } = {
+      request: {
+        headers: {},
+      },
+    };
+    // @ts-expect-error We mock subset of middleware inputs here.
+    await middlewareHandler(mock.fn(), {})(args);
+    assert.strictEqual(
+      args.request.headers['x-amz-user-agent'],
+      'testUserAgent'
+    );
   });
 });