+++ packages/ai-constructs/src/conversation/runtime/bedrock_converse_adapter.ts
@@ -7,6 +7,7 @@ import {
   Message,
   Tool,
   ToolConfiguration,
+  ToolInputSchema,
 } from '@aws-sdk/client-bedrock-runtime';
 import {
   ConversationTurnEvent,
@@ -40,8 +41,24 @@ export class BedrockConverseAdapter {
     eventToolsProvider = new ConversationTurnEventToolsProvider(event),
     private readonly messageHistoryRetriever = new ConversationMessageHistoryRetriever(
       event
-    )
+    ),
+    private readonly logger = console
   ) {
+    if (event.request.headers['x-amz-user-agent']) {
+      this.bedrockClient.middlewareStack.add(
+        (next) => (args) => {
+          // @ts-expect-error Request is typed as unknown.
+          // But this is recommended way to alter headers per https://github.com/aws/aws-sdk-js-v3/blob/main/README.md.
+          args.request.headers['x-amz-user-agent'] =
+            event.request.headers['x-amz-user-agent'];
+          return next(args);
+        },
+        {
+          step: 'build',
+          name: 'amplify-user-agent-injector',
+        }
+      );
+    }
     this.executableTools = [
       ...eventToolsProvider.getEventTools(),
       ...additionalTools,
@@ -90,9 +107,16 @@ export class BedrockConverseAdapter {
         inferenceConfig: inferenceConfiguration,
         toolConfig,
       };
+      this.logger.info('Sending Bedrock Converse request');
+      this.logger.debug('Bedrock Converse request:', converseCommandInput);
       bedrockResponse = await this.bedrockClient.send(
         new ConverseCommand(converseCommandInput)
       );
+      this.logger.info(
+        `Received Bedrock Converse response, requestId=${bedrockResponse.$metadata.requestId}`,
+        bedrockResponse.usage
+      );
+      this.logger.debug('Bedrock Converse response:', bedrockResponse);
       if (bedrockResponse.output?.message) {
         messages.push(bedrockResponse.output?.message);
       }
@@ -112,12 +136,17 @@ export class BedrockConverseAdapter {
           // and propagate result back to client.
           return clientToolUseBlocks;
         }
+        const toolResponseContentBlocks: Array<ContentBlock> = [];
         for (const responseContentBlock of toolUseBlocks) {
           const toolUseBlock =
             responseContentBlock as ContentBlock.ToolUseMember;
-          const toolMessage = await this.executeTool(toolUseBlock);
-          messages.push(toolMessage);
+          const toolResultContentBlock = await this.executeTool(toolUseBlock);
+          toolResponseContentBlocks.push(toolResultContentBlock);
         }
+        messages.push({
+          role: 'user',
+          content: toolResponseContentBlocks,
+        });
       }
     } while (bedrockResponse.stopReason === 'tool_use');
 
@@ -171,7 +200,9 @@ export class BedrockConverseAdapter {
           toolSpec: {
             name: t.name,
             description: t.description,
-            inputSchema: t.inputSchema,
+            // We have to cast to bedrock type as we're using different types to describe JSON schema in our API.
+            // These types are runtime compatible.
+            inputSchema: t.inputSchema as ToolInputSchema,
           },
         };
       }),
@@ -180,7 +211,7 @@ export class BedrockConverseAdapter {
 
   private executeTool = async (
     toolUseBlock: ContentBlock.ToolUseMember
-  ): Promise<Message> => {
+  ): Promise<ContentBlock> => {
     if (!toolUseBlock.toolUse.name) {
       throw Error('Bedrock tool use response is missing a tool name');
     }
@@ -191,45 +222,34 @@ export class BedrockConverseAdapter {
       );
     }
     try {
+      this.logger.info(`Invoking tool ${tool.name}`);
+      this.logger.debug('Tool input:', toolUseBlock.toolUse.input);
       const toolResponse = await tool.execute(toolUseBlock.toolUse.input);
+      this.logger.info(`Received response from ${tool.name} tool`);
+      this.logger.debug(toolResponse);
       return {
-        role: 'user',
-        content: [
-          {
-            toolResult: {
-              toolUseId: toolUseBlock.toolUse.toolUseId,
-              content: [toolResponse],
-              status: 'success',
-            },
-          },
-        ],
+        toolResult: {
+          toolUseId: toolUseBlock.toolUse.toolUseId,
+          content: [toolResponse],
+          status: 'success',
+        },
       };
     } catch (e) {
       if (e instanceof Error) {
         return {
-          role: 'user',
-          content: [
-            {
-              toolResult: {
-                toolUseId: toolUseBlock.toolUse.toolUseId,
-                content: [{ text: e.toString() }],
-                status: 'error',
-              },
-            },
-          ],
+          toolResult: {
+            toolUseId: toolUseBlock.toolUse.toolUseId,
+            content: [{ text: e.toString() }],
+            status: 'error',
+          },
         };
       }
       return {
-        role: 'user',
-        content: [
-          {
-            toolResult: {
-              toolUseId: toolUseBlock.toolUse.toolUseId,
-              content: [{ text: 'unknown error occurred' }],
-              status: 'error',
-            },
-          },
-        ],
+        toolResult: {
+          toolUseId: toolUseBlock.toolUse.toolUseId,
+          content: [{ text: 'unknown error occurred' }],
+          status: 'error',
+        },
       };
     }
   };