+++ packages/ai-constructs/src/conversation/runtime/conversation_turn_executor.test.ts
@@ -11,7 +11,6 @@ void describe('Conversation turn executor', () => {
     conversationId: 'testConversationId',
     currentMessageId: 'testCurrentMessageId',
     graphqlApiEndpoint: '',
-    messages: [],
     messageHistoryQuery: {
       getQueryName: '',
       getQueryInputTypeName: '',
@@ -47,9 +46,11 @@ void describe('Conversation turn executor', () => {
 
     const consoleErrorMock = mock.fn();
     const consoleLogMock = mock.fn();
+    const consoleDebugMock = mock.fn();
     const consoleMock = {
       error: consoleErrorMock,
       log: consoleLogMock,
+      debug: consoleDebugMock,
     } as unknown as Console;
 
     await new ConversationTurnExecutor(
@@ -100,9 +101,11 @@ void describe('Conversation turn executor', () => {
 
     const consoleErrorMock = mock.fn();
     const consoleLogMock = mock.fn();
+    const consoleDebugMock = mock.fn();
     const consoleMock = {
       error: consoleErrorMock,
       log: consoleLogMock,
+      debug: consoleDebugMock,
     } as unknown as Console;
 
     await assert.rejects(
@@ -164,9 +167,11 @@ void describe('Conversation turn executor', () => {
 
     const consoleErrorMock = mock.fn();
     const consoleLogMock = mock.fn();
+    const consoleDebugMock = mock.fn();
     const consoleMock = {
       error: consoleErrorMock,
       log: consoleLogMock,
+      debug: consoleDebugMock,
     } as unknown as Console;
 
     await assert.rejects(