+++ packages/ai-constructs/src/conversation/runtime/conversation_turn_response_sender.ts
@@ -22,19 +22,19 @@ export class ConversationTurnResponseSender {
     private readonly event: ConversationTurnEvent,
     private readonly graphqlRequestExecutor = new GraphqlRequestExecutor(
       event.graphqlApiEndpoint,
-      event.request.headers.authorization
-    )
+      event.request.headers.authorization,
+      event.request.headers['x-amz-user-agent']
+    ),
+    private readonly logger = console
   ) {}
 
   sendResponse = async (message: ContentBlock[]) => {
-    const { query, variables } = this.createMutationRequest(message);
+    const responseMutationRequest = this.createMutationRequest(message);
+    this.logger.debug('Sending response mutation:', responseMutationRequest);
     await this.graphqlRequestExecutor.executeGraphql<
       MutationResponseInput,
       void
-    >({
-      query,
-      variables,
-    });
+    >(responseMutationRequest);
   };
 
   private createMutationRequest = (content: ContentBlock[]) => {