+++ packages/sandbox/src/sandbox_singleton_factory.ts
@@ -14,7 +14,6 @@ import { LambdaClient } from '@aws-sdk/client-lambda';
 import { BackendOutputClientFactory } from '@aws-amplify/deployed-backend-client';
 import { LambdaFunctionLogStreamer } from './lambda_function_log_streamer.js';
 import { CloudWatchLogEventMonitor } from './cloudwatch_logs_monitor.js';
-import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
 
 /**
  * Factory to create a new sandbox
@@ -41,7 +40,6 @@ export class SandboxSingletonFactory {
         packageManagerControllerFactory.getPackageManagerController(),
         this.format
       );
-      const cfnClient = new CloudFormationClient();
       this.instance = new FileWatchingSandbox(
         this.sandboxIdResolver,
         new AmplifySandboxExecutor(
@@ -52,7 +50,6 @@ export class SandboxSingletonFactory {
         new SSMClient(),
         new LambdaFunctionLogStreamer(
           new LambdaClient(),
-          cfnClient,
           new CloudWatchLogEventMonitor(new CloudWatchLogsClient()),
           BackendOutputClientFactory.getInstance(),
           this.printer