+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @aws-amplify/backend
 
+## 1.4.0
+
+### Minor Changes
+
+- f5d0ab4: adds support to reference existing layers in defineFunction
+
+### Patch Changes
+
+- Updated dependencies [f5d0ab4]
+  - @aws-amplify/backend-function@1.6.0
+
+## 1.3.2
+
+### Patch Changes
+
+- 0a5e51c: Stream conversation logs in sandbox
+- Updated dependencies [0a5e51c]
+  - @aws-amplify/backend-output-schemas@1.3.0
+
 ## 1.3.1
 
 ### Patch Changes